secure_file_upload_demo/requirements.txt
Flask
python-magic-bin
